Refactor ScatterplotExample
