+++ examples/scatterplot/app.js
@@ -1,13 +1,6 @@
-/* global window,document */
-import React, {Component} from 'react';
-import {render} from 'react-dom';
-import MapGL from 'react-map-gl';
-import DeckGLOverlay from './deckgl-overlay.js';
-
+import {ScatterplotLayer} from 'deck.gl';
 import {json as requestJson} from 'd3-request';
 
-// Set your mapbox token here
-const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
 const MALE_COLOR = [0, 128, 255];
 const FEMALE_COLOR = [255, 0, 128];
 
@@ -15,62 +8,49 @@ const FEMALE_COLOR = [255, 0, 128];
 const DATA_URL =
   'https://raw.githubusercontent.com/uber-common/deck.gl-data/master/examples/scatterplot/manhattan.json'; // eslint-disable-line
 
-class Root extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      viewport: {
-        ...DeckGLOverlay.defaultViewport,
-        width: 500,
-        height: 500
-      },
-      data: null
-    };
+export default class App {
 
+  constructor({onUpdated = () => {}} = {}) {
+    this.data = null;
     requestJson(DATA_URL, (error, response) => {
       if (!error) {
-        this.setState({data: response});
+        this.data = response;
+        onUpdated();
       }
     });
   }
 
-  componentDidMount() {
-    window.addEventListener('resize', this._resize.bind(this));
-    this._resize();
+  getDefaultViewState() {
+    return {
+      longitude: -74,
+      latitude: 40.7,
+      zoom: 11,
+      maxZoom: 16,
+      pitch: 0,
+      bearing: 0
+    };
   }
 
-  _resize() {
-    this._onViewportChange({
-      width: window.innerWidth,
-      height: window.innerHeight
-    });
-  }
+  renderLayers(props) {
+    const {
+      maleColor = MALE_COLOR,
+      femaleColor = FEMALE_COLOR,
+      radius = 20
+    } = props;
 
-  _onViewportChange(viewport) {
-    this.setState({
-      viewport: {...this.state.viewport, ...viewport}
-    });
-  }
-
-  render() {
-    const {viewport, data} = this.state;
-
-    return (
-      <MapGL
-        {...viewport}
-        onViewportChange={this._onViewportChange.bind(this)}
-        mapboxApiAccessToken={MAPBOX_TOKEN}
-      >
-        <DeckGLOverlay
-          viewport={viewport}
-          data={data}
-          maleColor={MALE_COLOR}
-          femaleColor={FEMALE_COLOR}
-          radius={30}
-        />
-      </MapGL>
-    );
+    return [
+      this.data && new ScatterplotLayer({
+        id: 'scatter-plot',
+        data: this.data,
+        radiusScale: radius,
+        radiusMinPixels: 0.25,
+        getPosition: d => [d[0], d[1], 0],
+        getColor: d => (d[2] === 1 ? maleColor : femaleColor),
+        getRadius: d => 1,
+        updateTriggers: {
+          getColor: [maleColor, femaleColor]
+        }
+      })
+    ];
   }
 }
-
-render(<Root />, document.body.appendChild(document.createElement('div')));
