+++ examples/scatterplot/webpack.config.js
@@ -7,7 +7,7 @@ const webpack = require('webpack');
 
 const CONFIG = {
   entry: {
-    app: resolve('./app.js')
+    app: resolve('./react.js')
   },
 
   devtool: 'source-map',
